setting up pip install
